fix(edit-user): remove has wrong search criteria and history had uniqu column

+++ backend/src/common/base/base-dao.class.ts
@@ -73,9 +73,9 @@ export abstract class BaseDAO<T> {
     return this.repository.softDelete(findCriteria);
   }
 
-  async remove(findCriteria: FindOptionsWhere<T>) {
+  async remove(findCriteria: FindOneOptions<T>) {
     const record = await this.get(findCriteria);
-    return this.repository.remove(record);
+    if (record) return this.repository.remove(record);
   }
 
   public async getManyPaginated(
